requirements.txt
google-generativeai
prodiapy
discord
flask
python_dotenv